implemented workaround for the camera preview
The surface's callback isn't called when the phone's screen gets off, however onPause is called and the camera is stopped, which is correct. But when onResume is called and the camera started again the surface view's callback isn't called therefore startPreview isn't called as well and the camera and preview doesn't work anymore.